% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redrake-package.R
\docType{package}
\name{redrake-package}
\alias{redrake}
\alias{redrake-package}
\title{redrake: Simple Reproducible Workflows}
\description{
A simpler interface to drake for R. Includes a converter for remake projects.
}
\details{
Work in progress.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/krlmlr/redrake}
  \item \url{http://krlmlr.github.io/redrake}
  \item Report bugs at \url{https://github.com/krlmlr/redrake/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

}
